Refactor: Move and rename BlazeParentActivity to blazepromote package
