+++ WordPress/src/main/AndroidManifest.xml
@@ -1065,7 +1065,7 @@
             android:name="com.google.mlkit.vision.DEPENDENCIES"
             android:value="barcode_ui"/>
 
-        <activity android:name=".ui.blaze.BlazeParentActivity"
+        <activity android:name=".ui.blaze.blazepromote.BlazeParentActivity"
             android:exported="false"
             android:label="@string/blaze_activity_title"
             android:screenOrientation="portrait"
